fixed answer string bug
