+++ src/main/java/dev/salonce/discordquizbot/application/ButtonHandler.java
@@ -1,4 +1,4 @@
-package dev.salonce.discordquizbot.infrastructure.buttons;
+package dev.salonce.discordquizbot.application;
 
 import dev.salonce.discordquizbot.infrastructure.dtos.ButtonInteractionData;
 import discord4j.core.event.domain.interaction.ButtonInteractionEvent;
