+++ src/main/java/dev/salonce/discordquizbot/infrastructure/buttons/handlers/LeaveMatchButtonHandler.java
@@ -1,6 +1,6 @@
 package dev.salonce.discordquizbot.infrastructure.buttons.handlers;
 
-import dev.salonce.discordquizbot.infrastructure.buttons.ButtonHandler;
+import dev.salonce.discordquizbot.application.ButtonHandler;
 import dev.salonce.discordquizbot.infrastructure.dtos.ButtonInteractionData;
 import dev.salonce.discordquizbot.application.MatchService;
 import discord4j.core.event.domain.interaction.ButtonInteractionEvent;
